+++ src/components/button-show-more.js
@@ -1,0 +1,3 @@
+const getButtonShowMore = () => `<button class="films-list__show-more">Show more</button>`;
+
+export default getButtonShowMore;
