+++ src/components/movie-main-container.js
@@ -1,0 +1,3 @@
+const getMovieContainer = () => `<section class="films"></section>`;
+
+export default getMovieContainer;
